improve modification observers
