+++ src/methods/addModificationObservers.js
@@ -1,20 +1,41 @@
 // @ts-check
 import '../types'
 
+import { removeModificationObservers } from './'
+
 /**
- * Adds event-listeners to the selectorArea
- * @param {DSArea[]} nodes
- * @param {DSModificationCallback} callback
- * @param {MutationObserver} modificationObserver
+ * Removes event-listeners from the DOMNode
+ * @typedef {()=>void} DSCleanup
  */
-export default (nodes, callback, modificationObserver) => {
+
+/**
+ * Adds event-listeners to a DOMNode
+ * @param {DSArea[]} nodes
+ * @param {DSModificationCallback} cb
+ * @return {{observer:MutationObserver,callback:DSModificationCallback,cleanup:DSCleanup}}
+ */
+export default (nodes, cb) => {
+  const callback = cb
+
   window.addEventListener('resize', callback)
   window.addEventListener('scroll', callback)
 
+  const observer = new MutationObserver(callback)
+
   nodes.forEach((el, i) => {
-    modificationObserver.observe(el, {
+    observer.observe(el, {
       childList: i !== 0,
       attributes: true,
     })
   })
+
+  /**
+   * Removes all observers
+   */
+  const cleanup = () => removeModificationObservers(
+    observer,
+    callback,
+  )
+
+  return { observer, callback, cleanup }
 }
