+++ src/methods/index.js
@@ -5,6 +5,7 @@ export { default as createSelectorAreaElement } from './createSelectorAreaElemen
 export { default as createSelectorElement } from './createSelectorElement'
 export { default as debounce } from './debounce'
 export { default as documentScroll } from './documentScroll'
+export { default as getAllParentNodes } from './getAllParentNodes'
 export { default as getAreaRect } from './getAreaRect'
 export { default as getCurrentScroll } from './getCurrentScroll'
 export { default as getOverflowEdges } from './getOverflowEdges'
